Update uniforms.xml
Added descriptions for all uniforms that have special pockets and describe what they do. All storage_clothing is also a mediumitem now so it cannot fit in a exosuit or container and must be carried in hand. Small adjustments in line with vanilla armor protection values for suits like the prisoner which has civilian clothing protection.